update basis module
