+++ src/horton_part/data/gauss.json
@@ -1,0 +1,280 @@
+{
+  "1": [
+    [
+      2.0,
+      2.0,
+      2.0,
+      2.0
+    ],
+    [
+      5.672,
+      1.505,
+      0.5308,
+      0.2204
+    ],
+    [
+      0.04588955,
+      0.26113177,
+      0.47966863,
+      0.21331515
+    ]
+  ],
+  "3": [
+    [
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0
+    ],
+    [
+      60.3528,
+      14.895,
+      5.0545,
+      1.9759,
+      0.0971,
+      0.0314
+    ],
+    [
+      0.05454438,
+      0.35593951,
+      0.91132803,
+      0.62408593,
+      1.01350763,
+      0.04059578
+    ]
+  ],
+  "6": [
+    [
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0
+    ],
+    [
+      148.3,
+      42.19,
+      15.33,
+      6.146,
+      0.7846,
+      0.2511
+    ],
+    [
+      0.14213932,
+      0.58600661,
+      1.07985999,
+      0.0193642,
+      2.73124559,
+      1.44130948
+    ]
+  ],
+  "7": [
+    [
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0
+    ],
+    [
+      178.0,
+      52.42,
+      19.87,
+      1.276,
+      0.6291,
+      0.2857
+    ],
+    [
+      0.1762005,
+      0.64409882,
+      1.00359925,
+      2.34863505,
+      1.86405143,
+      0.96339936
+    ]
+  ],
+  "8": [
+    [
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0
+    ],
+    [
+      220.1,
+      65.66,
+      25.98,
+      1.685,
+      0.686,
+      0.2311
+    ],
+    [
+      0.20051893,
+      0.64160829,
+      0.98585628,
+      3.01184504,
+      2.70306065,
+      0.45711097
+    ]
+  ],
+  "9": [
+    [
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0
+    ],
+    [
+      319.6233,
+      85.5603,
+      31.7878,
+      2.4347,
+      1.0167,
+      0.3276
+    ],
+    [
+      0.16408974,
+      0.72344762,
+      0.93390509,
+      2.85966001,
+      3.51408367,
+      0.80481333
+    ]
+  ],
+  "14": [
+    [
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0
+    ],
+    [
+      4.0274,
+      0.4191,
+      8.3302,
+      88.4554,
+      0.076,
+      213.778,
+      0.2271,
+      592.0697
+    ],
+    [
+      4.66493968,
+      1.43077243,
+      3.02335498,
+      0.92989328,
+      0.29066472,
+      0.57081524,
+      2.8537414,
+      0.23581856
+    ]
+  ],
+  "16": [
+    [
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0
+    ],
+    [
+      715.0735,
+      240.4533,
+      120.1752,
+      14.3796,
+      7.0821,
+      0.5548,
+      0.5176,
+      0.2499,
+      0.1035
+    ],
+    [
+      0.29455963,
+      0.58698056,
+      0.80997645,
+      1.57119984,
+      5.92511019,
+      2.30623427,
+      2.25300903,
+      2.07579989,
+      0.17713248
+    ]
+  ],
+  "17": [
+    [
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0
+    ],
+    [
+      1139.4058,
+      379.1381,
+      0.1635,
+      0.6579,
+      0.3952,
+      19.5054,
+      8.9484,
+      151.8297
+    ],
+    [
+      0.16005908,
+      0.49615207,
+      0.67085328,
+      4.54949229,
+      2.68454233,
+      0.93361892,
+      6.49142386,
+      1.0138586
+    ]
+  ],
+  "35": [
+    [
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0,
+      2.0
+    ],
+    [
+      0.1136,
+      5.2615,
+      0.2747,
+      1026.4314,
+      1.6828,
+      0.5514,
+      64.8824
+    ],
+    [
+      0.26002704,
+      17.04051595,
+      3.02612546,
+      1.54627807,
+      1.25096851,
+      5.26425613,
+      6.61182792
+    ]
+  ]
+}
