--- src/horton_part/data/gauss_exponents.json
@@ -1,88 +1,0 @@
-{
-  "1": [
-    5.672,
-    1.505,
-    0.5308,
-    0.2204
-  ],
-  "3": [
-    60.3528,
-    14.895,
-    5.0545,
-    1.9759,
-    0.0971,
-    0.0314
-  ],
-  "6": [
-    148.3,
-    42.19,
-    15.33,
-    6.146,
-    0.7846,
-    0.2511
-  ],
-  "7": [
-    178.0,
-    52.42,
-    19.87,
-    1.276,
-    0.6291,
-    0.2857
-  ],
-  "8": [
-    220.1,
-    65.66,
-    25.98,
-    1.685,
-    0.686,
-    0.2311
-  ],
-  "9": [
-    319.6233,
-    85.5603,
-    31.7878,
-    2.4347,
-    1.0167,
-    0.3276
-  ],
-  "14": [
-    4.0274,
-    0.4191,
-    8.3302,
-    88.4554,
-    0.076,
-    213.778,
-    0.2271,
-    592.0697
-  ],
-  "16": [
-    715.0735,
-    240.4533,
-    120.1752,
-    14.3796,
-    7.0821,
-    0.5548,
-    0.5176,
-    0.2499,
-    0.1035
-  ],
-  "17": [
-    1139.4058,
-    379.1381,
-    0.1635,
-    0.6579,
-    0.3952,
-    19.5054,
-    8.9484,
-    151.8297
-  ],
-  "35": [
-    0.1136,
-    5.2615,
-    0.2747,
-    1026.4314,
-    1.6828,
-    0.5514,
-    64.8824
-  ]
-}
