--- src/horton_part/data/gauss_initials.json
@@ -1,72 +1,0 @@
-{
-  "1": [
-    0.04588955,
-    0.26113177,
-    0.47966863,
-    0.21331515
-  ],
-  "6": [
-    0.14213932,
-    0.58600661,
-    1.07985999,
-    0.0193642,
-    2.73124559,
-    1.44130948
-  ],
-  "7": [
-    0.1762005,
-    0.64409882,
-    1.00359925,
-    2.34863505,
-    1.86405143,
-    0.96339936
-  ],
-  "8": [
-    0.20051893,
-    0.64160829,
-    0.98585628,
-    3.01184504,
-    2.70306065,
-    0.45711097
-  ],
-  "14": [
-    4.66493968,
-    1.43077243,
-    3.02335498,
-    0.92989328,
-    0.29066472,
-    0.57081524,
-    2.8537414,
-    0.23581856
-  ],
-  "16": [
-    0.29455963,
-    0.58698056,
-    0.80997645,
-    1.57119984,
-    5.92511019,
-    2.30623427,
-    2.25300903,
-    2.07579989,
-    0.17713248
-  ],
-  "17": [
-    0.16005908,
-    0.49615207,
-    0.67085328,
-    4.54949229,
-    2.68454233,
-    0.93361892,
-    6.49142386,
-    1.0138586
-  ],
-  "35": [
-    0.26002704,
-    17.04051595,
-    3.02612546,
-    1.54627807,
-    1.25096851,
-    5.26425613,
-    6.61182792
-  ]
-}
