--- src/horton_part/data/slater_exponents.json
@@ -1,121 +1,0 @@
-{
-  "1": [
-    6.6,
-    4.62,
-    3.23,
-    2.26,
-    1.58,
-    1.1,
-    0.77,
-    1.0
-  ],
-  "6": [
-    22.8,
-    17.35,
-    13.2,
-    8.79,
-    10.83,
-    7.91,
-    5.78,
-    4.22,
-    4.51,
-    3.46,
-    2.65,
-    2.03,
-    1.56
-  ],
-  "7": [
-    25.7,
-    19.83,
-    15.3,
-    10.21,
-    12.58,
-    9.2,
-    6.72,
-    4.92,
-    5.26,
-    4.03,
-    3.1,
-    2.38,
-    1.82,
-    1.4
-  ],
-  "8": [
-    29.6,
-    22.76,
-    17.5,
-    11.58,
-    14.16,
-    10.29,
-    7.48,
-    5.43,
-    5.78,
-    4.41,
-    3.37,
-    2.57,
-    1.96,
-    1.5
-  ],
-  "9": [
-    34.6,
-    26.37,
-    20.1,
-    13.37,
-    16.44,
-    12.0,
-    8.75,
-    6.39,
-    6.81,
-    5.22,
-    4.0,
-    3.07,
-    2.35,
-    1.8
-  ],
-  "16": [
-    110.7,
-    84.96,
-    65.2,
-    50.04,
-    38.4,
-    24.21,
-    28.2,
-    19.74,
-    20.24,
-    14.97,
-    11.08,
-    10.82,
-    8.3,
-    6.36,
-    4.88,
-    4.65,
-    3.66,
-    2.88,
-    2.26,
-    1.78,
-    1.4
-  ],
-  "17": [
-    128.0,
-    96.41,
-    72.62,
-    54.7,
-    41.2,
-    25.97,
-    30.25,
-    21.17,
-    21.71,
-    16.06,
-    11.88,
-    11.6,
-    8.89,
-    6.82,
-    5.23,
-    4.99,
-    3.92,
-    3.08,
-    2.43,
-    1.91,
-    1.5
-  ]
-}
