--- src/horton_part/data/slater_initials.json
@@ -1,1 +1,0 @@
-{}
