+++ src/horton_part/utils.py
@@ -21,6 +21,7 @@
 """Utility Functions"""
 
 import numpy as np
+import warnings
 
 __all__ = [
     "typecheck_geo",
@@ -28,9 +29,18 @@ __all__ = [
     "radius_covalent",
     "wpart_schemes",
     "angstrom",
+    "compute_quantities",
+    "check_pro_atom_parameters",
+    "check_for_pro_error",
+    "check_for_grad_error",
+    "check_for_hessian_error",
+    "NEGATIVE_CUTOFF",
+    "POPULATION_CUTOFF",
 ]
 
 
+NEGATIVE_CUTOFF = -1e-12
+POPULATION_CUTOFF = 1e-4
 angstrom = 1.0e-10 / 0.5291772083e-10
 
 
@@ -409,3 +419,156 @@ radius_covalent = {
 radius_covalent = dict(
     [(k, v if v is None else v * angstrom) for k, v in list(radius_covalent.items())]
 )
+
+
+def compute_quantities(density, pro_atom_params, basis_functions, density_cutoff):
+    """
+    Compute various quantities based on input density, pro-atom parameters, and basis functions.
+
+    Parameters
+    ----------
+    density : ndarray
+        The density array.
+    pro_atom_params : ndarray
+        Array of pro-atom parameters.
+    basis_functions : ndarray
+        Array of basis functions.
+    density_cutoff : float, optional
+        The cutoff value for density and pro-atom density, below which values are considered negligible.
+
+    Returns
+    -------
+    pro_shells : ndarray
+        Calculated pro-atom shell values.
+    pro_density : ndarray
+        Calculated pro-atom density.
+    sick : ndarray
+        Boolean array indicating where density or pro-density is below the cutoff.
+    ratio : ndarray
+        Ratio of density to pro-atom density, with adjustments for sick values.
+    ln_ratio : ndarray
+        Natural logarithm of the ratio, with adjustments for sick values.
+    """
+    pro_atom_params = np.asarray(pro_atom_params).flatten()
+    pro_shells = basis_functions * pro_atom_params[:, None]
+    pro_density = np.einsum("ij->j", pro_shells)
+    sick = (density < density_cutoff) | (pro_density < density_cutoff)
+    with np.errstate(all="ignore"):
+        ratio = np.divide(density, pro_density, out=np.zeros_like(density), where=~sick)
+        ln_ratio = np.log(ratio, out=np.zeros_like(density), where=~sick)
+    return pro_shells, pro_density, sick, ratio, ln_ratio
+
+
+def check_pro_atom_parameters(
+    pro_atom_params,
+    basis_functions=None,
+    total_population=None,
+    pro_atom_density=None,
+    check_monotonicity=True,
+):
+    """
+    Check the validity of pro-atom parameters.
+
+    This function warns if any pro-atom parameters are negative and raises a
+    RuntimeError if the pro-atom density contains negative values. It also warns
+    if the sum of pro-atom parameters does not match the expected total atomic
+    population.
+
+    Parameters
+    ----------
+    pro_atom_params : ndarray
+        Array of pro-atom parameters.
+    basis_functions : ndarray, optional
+        Basis functions on the grid.
+    total_population : float, optional
+        Total atomic population to compare against the sum of pro-atom parameters.
+    pro_atom_density : ndarray, optional
+        Pre-calculated pro-atom density array. If not provided, it is calculated
+        from pro_atom_params and basis_functions.
+    check_monotonicity: bool, optional
+        Check if the density is monotonically decreased w.r.t. radial radius.
+
+    Raises
+    ------
+    RuntimeError
+        If negative values are found in the pro-atom density.
+
+    Warnings
+    --------
+    UserWarning
+        If any pro-atom parameters are negative.
+        If the sum of pro-atom parameters does not match the total_population.
+
+    Returns
+    -------
+    None
+    """
+    # Validate inputs
+    pro_atom_params = np.asarray(pro_atom_params)
+    if pro_atom_params.ndim != 1:
+        raise ValueError("pro_atom_params must be a 1D array")
+
+    if basis_functions is not None:
+        basis_functions = np.asarray(basis_functions)
+        if basis_functions.ndim != 2:
+            raise ValueError("basis_functions must be a 2D array")
+
+    if pro_atom_density is not None:
+        pro_atom_density = np.asarray(pro_atom_density)
+        if pro_atom_density.ndim != 1:
+            raise ValueError("pro_atom_density must be a 1D array")
+
+    # Check if pro-atom parameters are positive
+    if (pro_atom_params < NEGATIVE_CUTOFF).any():
+        warnings.warn("Not all pro-atom parameters are positive!")
+
+    # Calculate pro-atom density if not provided
+    if basis_functions is not None and pro_atom_density is None:
+        if pro_atom_params.size != basis_functions.shape[0]:
+            raise ValueError(
+                "Length of pro_atom_params does not match the number of basis functions"
+            )
+        pro_atom_density = (basis_functions * pro_atom_params[:, None]).sum(axis=0)
+
+    # Check for negative pro-atom density
+    if pro_atom_density is not None and (pro_atom_density < NEGATIVE_CUTOFF).any():
+        raise RuntimeError("Negative pro-atom density found!")
+
+    # Check if the sum of pro-atom parameters matches total population
+    if total_population is not None and not np.allclose(
+        np.sum(pro_atom_params), total_population, atol=POPULATION_CUTOFF
+    ):
+        warnings.warn(
+            "The sum of pro-atom parameters is not equal to atomic population"
+        )
+
+    if check_monotonicity and pro_atom_density is not None:
+        if (pro_atom_density[:-1] - pro_atom_density[1:] < NEGATIVE_CUTOFF).any():
+            raise RuntimeError("Pro-atom density should be monotonically decreasing.")
+
+
+def check_for_pro_error(pro, as_warn=True):
+    """Check for non-monotonic and non-negative density"""
+    if (pro < NEGATIVE_CUTOFF).any():
+        if as_warn:
+            warnings.warn("Negative pro-atom density found during optimization!")
+        else:
+            raise RuntimeError("Negative pro-atom density found!")
+    if (pro[:-1] - pro[1:] < NEGATIVE_CUTOFF).any():
+        if as_warn:
+            warnings.warn("Pro-atom density should be monotonically decreasing.")
+        else:
+            raise RuntimeError("Pro-atom density should be monotonically decreasing.")
+
+
+def check_for_grad_error(grad):
+    """Check for non-monotonic density and negative gradient errors."""
+    if (grad < NEGATIVE_CUTOFF).any():
+        raise RuntimeError("Negative gradient detected.")
+
+
+def check_for_hessian_error(hess):
+    """Check for negative-eigenvalue hessian errors."""
+    if (np.linalg.eigvals(hess) <= NEGATIVE_CUTOFF).any():
+        # raise RuntimeError("All eigenvalues of Hessian matrix are not all")
+        warnings.warn("All eigenvalues of Hessian matrix are not all")
